Move ExceptionOps from ledger-service/utils to //libs-scala/scala-utils
