+++ ledger-service/http-json-cli/BUILD.bazel
@@ -21,9 +21,9 @@ da_scala_library(
     visibility = ["//ledger-service:__subpackages__"],
     deps = [
         "//ledger-service/cli-opts",
-        "//ledger-service/utils",
         "//ledger/ledger-api-common",
         "//ledger/metrics",
+        "//libs-scala/scala-utils",
         "@maven//:ch_qos_logback_logback_classic",
         "@maven//:io_netty_netty_handler",
         "@maven//:org_slf4j_slf4j_api",
