+++ libs-scala/scala-utils/src/main/scala/com/daml/scalautil/ExceptionOps.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2021 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.util
+package com.daml.scalautil
 
 object ExceptionOps {
